lampo: reworking the handling code
