+++ lampo-async-jsonrpc/src/lib.rs
@@ -3,7 +3,8 @@
 use std::future::Future;
 use std::sync::Arc;
 
-use jsonrpsee::server::{RpcModule, RpcServiceBuilder, Server};
+use jsonrpsee::server::{RpcModule, RpcServiceBuilder, Server, ServerHandle};
+use tokio::runtime::Runtime;
 
 pub use jsonrpsee::types::{ErrorObject, ResponsePayload};
 pub use jsonrpsee::IntoResponse;
@@ -11,12 +12,14 @@ pub use jsonrpsee::IntoResponse;
 /// JSONRPC v2
 pub struct JSONRPCv2<T: Sync + Send + 'static> {
     inner: RpcModule<Arc<T>>,
+    host: String,
 }
 
 impl<T: Sync + Send + 'static> JSONRPCv2<T> {
-    pub fn new(ctx: Arc<T>, path: &str) -> anyhow::Result<Self> {
+    pub fn new(ctx: Arc<T>, host: &str) -> anyhow::Result<Self> {
         Ok(Self {
             inner: RpcModule::new(ctx),
+            host: host.to_owned(),
         })
     }
 
@@ -26,7 +29,6 @@ impl<T: Sync + Send + 'static> JSONRPCv2<T> {
         Fut: Future<Output = R> + Send,
         Fun: (Fn(Arc<T>, serde_json::Value) -> Fut) + Clone + Send + Sync + 'static,
     {
-        // FIXME: fix the type definition under here to avoid Arc<Arc< T>>
         self.inner.register_async_method(name, move |params, ctx| {
             let request: serde_json::Value = params.parse().unwrap();
             callback(ctx.as_ref().clone(), request)
@@ -34,16 +36,28 @@ impl<T: Sync + Send + 'static> JSONRPCv2<T> {
         Ok(())
     }
 
-    pub async fn listen(self) -> std::io::Result<()> {
+    pub async fn listen(self) -> std::io::Result<ServerHandle> {
         let rpc_middleware = RpcServiceBuilder::new().rpc_logger(1024);
         let server = Server::builder()
             .set_rpc_middleware(rpc_middleware)
-            .build("127.0.0.1:0")
+            .build(self.host)
             .await?;
-        let _addr = server.local_addr()?;
         let handle = server.start(self.inner);
-        // FIXME: stop the server in a proprer way
-        tokio::spawn(handle.stopped());
+        tokio::spawn(handle.clone().stopped());
+        Ok(handle)
+    }
+
+    /// Spawing the JSON RPC server on a new thread and a
+    /// personal runtime to handle specific RPC call.
+    pub fn spawn(self) -> anyhow::Result<()> {
+        std::thread::spawn(move || {
+            // We should create a single runtime for the JSON RPC server.
+            let rt = Runtime::new().unwrap();
+            rt.spawn(self.listen())
+        });
+        // FIXME: return the handler, so we should use a channel a some point.
         Ok(())
     }
+
+    // FIXME: add `spawn_with_runtime` if necessary.
 }
