+++ lampod/src/actions/handler.rs
@@ -7,8 +7,10 @@ use lampo_common::error;
 use lampo_common::error::Ok;
 use lampo_common::event::ln::LightningEvent;
 use lampo_common::event::{Emitter, Event, Subscriber};
+use lampo_common::handler::ExternalHandler;
 use lampo_common::handler::Handler as EventHandler;
 use lampo_common::json;
+use lampo_common::jsonrpc::Request;
 use lampo_common::ldk;
 use lampo_common::model::response::PaymentHop;
 use lampo_common::model::response::PaymentState;
@@ -16,13 +18,11 @@ use lampo_common::types::ChannelState;
 
 use crate::chain::{LampoChainManager, WalletManager};
 use crate::command::Command;
-use crate::handler::external_handler::ExternalHandler;
-use crate::json_rpc2::Request;
 use crate::ln::events::PeerEvents;
 use crate::ln::{LampoChannelManager, LampoInventoryManager, LampoPeerManager};
-use crate::{async_run, LampoDaemon};
+use crate::LampoDaemon;
 
-use super::{Handler, InventoryHandler};
+use super::Handler;
 
 pub struct LampoHandler {
     channel_manager: Arc<LampoChannelManager>,
@@ -95,22 +95,11 @@ impl EventHandler for LampoHandler {
 #[allow(unused_variables)]
 impl Handler for LampoHandler {
     fn react(&self, event: crate::command::Command) -> error::Result<()> {
+        let handler = self.external_handlers.borrow();
         match event {
-            Command::LNCommand => unimplemented!(),
-            Command::OnChainCommand => unimplemented!(),
-            Command::PeerEvent(event) => {
-                async_run!(self.peer_manager.handle(event))
-            }
-            Command::InventoryEvent(event) => {
-                self.inventory_manager.handle(event)?;
-                Ok(())
-            }
             Command::ExternalCommand(req, chan) => {
-                log::info!(
-                    "external handler size {}",
-                    self.external_handlers.borrow().len()
-                );
-                for handler in self.external_handlers.borrow().iter() {
+                log::debug!(target: "lampo", "external handler size {}", handler.len());
+                for handler in handler.iter() {
                     if let Some(resp) = handler.handle(&req)? {
                         chan.send(resp)?;
                         return Ok(());
