--- lampod/src/handler.rs
@@ -1,2 +1,0 @@
-//! Lampo Handler module implementation.
-pub mod external_handler;
