+++ lampod/src/jsonrpc/channels.rs
@@ -2,11 +2,12 @@ use lampo_common::event::ln::LightningEvent;
 use lampo_common::event::Event;
 use lampo_common::handler::Handler;
 use lampo_common::json;
+use lampo_common::jsonrpc::{Error, RpcError};
 use lampo_common::model::request;
 use lampo_common::model::response;
 
-use crate::json_rpc2::{Error, RpcError};
 use crate::ln::events::ChannelEvents;
+use crate::rpc_error;
 use crate::LampoDaemon;
 
 pub fn json_list_channels(ctx: &LampoDaemon, request: &json::Value) -> Result<json::Value, Error> {
