+++ lampod/src/jsonrpc/offchain.rs
@@ -6,6 +6,7 @@ use lampo_common::conf::Network;
 use lampo_common::event::ln::LightningEvent;
 use lampo_common::event::Event;
 use lampo_common::handler::Handler;
+use lampo_common::jsonrpc::{Error, RpcError};
 use lampo_common::ldk;
 use lampo_common::ldk::offers::offer;
 use lampo_common::ldk::offers::offer::Amount;
@@ -18,7 +19,6 @@ use lampo_common::model::response::PayResult;
 use lampo_common::model::response::{Invoice, InvoiceInfo};
 use lampo_common::{json, model::request::DecodeInvoice};
 
-use crate::json_rpc2::{Error, RpcError};
 use crate::rpc_error;
 use crate::LampoDaemon;
