+++ lampod/src/jsonrpc/onchain.rs
@@ -1,7 +1,7 @@
 //! On Chain RPC methods
 use lampo_common::json;
+use lampo_common::jsonrpc::{Error, RpcError};
 
-use crate::json_rpc2::{Error, RpcError};
 use crate::LampoDaemon;
 
 pub fn json_new_addr(ctx: &LampoDaemon, request: &json::Value) -> Result<json::Value, Error> {
