+++ lampod/src/jsonrpc/peer_control.rs
@@ -1,9 +1,8 @@
 //! Peer Control JSON RPC Interface!
 use lampo_common::json;
+use lampo_common::jsonrpc::{Error, RpcError};
 use lampo_common::model::Connect;
 
-use crate::json_rpc2::Error;
-use crate::json_rpc2::RpcError;
 use crate::{ln::events::PeerEvents, LampoDaemon};
 
 pub fn json_connect(ctx: &LampoDaemon, request: &json::Value) -> Result<json::Value, Error> {
